server/requirements.txt
django
paho-mqtt
django-bootstrap-v5